Started to use doubles for layout and painting
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk/qwt@661 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ src/qwt_round_scale_draw.cpp
@@ -200,7 +200,7 @@ void QwtRoundScaleDraw::drawLabel(QPainter *painter, double value) const
 
    \sa drawBackbone(), drawLabel() 
 */
-void QwtRoundScaleDraw::drawTick(QPainter *painter, double value, int len) const
+void QwtRoundScaleDraw::drawTick(QPainter *painter, double value, double len) const
 {
     if ( len <= 0 )
         return;
